Fix bug where player can jump again, after a successful hit, and land on the other side of the net. At the start of the next play, the player would then try to run all the way back.
To things help with the issue. The first, is that if the player is on the other side of the next when a match restarts, he teleports (as he does when we have a game restart).

We also don't end the match until the player is on the ground and in idle. That way we don't get any weird cases where the player is in the air still when the serve comes in.