Found that the MusicManager was a bit of a performance hit. The issue seems to be that calls to query the MediaPlayer state are expensive. MusicManager has been updated to be more event based, to avoid querying the MediaPlayer every frame.
The only down side is that there doesn't seem to be an event when the game regains control of the media player, so if the user pauses their own music mid game, the game music won't start up right away. It will wait until the nect state change. Seems minor compared to the major performance hit required to query every frame.