Fix 2 issues with leaderboards.
The first was simply that Trick Attack was not being updated, unless the user opened the Google Leaderboard Actitivty. It now updates the moment a new highscore is achieved.

The second issue appears to be an issues with the Google Play Services API. If I submit multiple scores in a short peroid (maybe between syncing with the server), the score that appears on the google leaderboard is not the best score, but the most recently submitted. To fix this I now only submit the score if it is better than the high score.