Tap events on Android and Windows now fire on PRESS rather than RELEASE.
This is to make hitting easier, which feels very delayed because it happens on release. 

Unfortunetly this change is isolated to hitting; it affects all Tap gestures in the game (such as menus). This will be problematic later on if we ever add sliding menus or anything that relys on Tap and Swipe being mutually exclusive. Right now that doesnt exist, so hacking it in.